Move lock check to entrypoint
